refactor(exceptions) : 체스 게임 내부 로직 위반을 통한 예외는 ChessException을 통해, 방 생성 예외는 RoomException을 통해 처리
