+++ src/main/java/chess/controller/ChessController.java
@@ -4,7 +4,7 @@ import chess.domain.ChessGame;
 import chess.domain.board.Board;
 import chess.domain.command.Command;
 import chess.domain.position.Position;
-import chess.exception.ChessException;
+import chess.exception.chessgame.ChessException;
 import chess.view.InputView;
 import chess.view.OutputView;
