+++ src/main/java/chess/domain/piece/Blank.java
@@ -2,7 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.Side;
 import chess.domain.position.Position;
-import chess.exception.PieceDoesNotExistException;
+import chess.exception.chessgame.PieceDoesNotExistException;
 
 import java.util.List;
