+++ src/main/java/chess/exception/chessgame/ChessException.java
@@ -1,4 +1,4 @@
-package chess.exception;
+package chess.exception.chessgame;
 
 public class ChessException extends RuntimeException {
     public ChessException(String message) {
