+++ src/main/java/chess/exception/chessgame/InvalidMethodCallException.java
@@ -1,4 +1,4 @@
-package chess.exception;
+package chess.exception.chessgame;
 
 public class InvalidMethodCallException extends UnsupportedOperationException {
     private static final String INVALID_METHOD_CALL_MESSAGE = "유효하지 않은 메서드 호출입니다.";
