+++ src/main/java/chess/exception/chessgame/InvalidMovementException.java
@@ -1,4 +1,4 @@
-package chess.exception;
+package chess.exception.chessgame;
 
 public class InvalidMovementException extends ChessException {
     public InvalidMovementException(String message) {
