+++ src/main/java/chess/exception/chessgame/PieceDoesNotExistException.java
@@ -1,4 +1,4 @@
-package chess.exception;
+package chess.exception.chessgame;
 
 public class PieceDoesNotExistException extends ChessException {
