+++ src/main/java/chess/exception/room/DuplicateRoomNameException.java
@@ -1,0 +1,8 @@
+package chess.exception.room;
+
+public class DuplicateRoomNameException extends RoomException {
+
+    public DuplicateRoomNameException() {
+        super("해당 이름의 방이 이미 존재합니다.");
+    }
+}
