+++ src/main/java/chess/exception/room/NotExistRoomException.java
@@ -1,6 +1,6 @@
-package chess.exception;
+package chess.exception.room;
 
-public class NotExistRoomException extends ChessException {
+public class NotExistRoomException extends RoomException {
 
     public NotExistRoomException() {
         super("방이 존재하지 않습니다.");
