+++ src/main/java/chess/exception/room/RoomException.java
@@ -1,0 +1,10 @@
+package chess.exception.room;
+
+import org.springframework.dao.DataAccessException;
+
+public class RoomException extends DataAccessException {
+
+    public RoomException(String msg) {
+        super(msg);
+    }
+}
