+++ src/test/java/chess/domain/piece/BlankTest.java
@@ -1,7 +1,7 @@
 package chess.domain.piece;
 
 import chess.domain.position.Position;
-import chess.exception.PieceDoesNotExistException;
+import chess.exception.chessgame.PieceDoesNotExistException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
